Fix libusb_get_device_list return value
It was documented to return the list length, but was returning 0.

+++ libusb/libusb.h
@@ -635,7 +635,7 @@ struct libusb_transfer {
 int libusb_init(void);
 void libusb_exit(void);
 
-int libusb_get_device_list(libusb_device ***list);
+size_t libusb_get_device_list(libusb_device ***list);
 void libusb_free_device_list(libusb_device **list, int unref_devices);
 const struct libusb_device_descriptor *libusb_get_device_descriptor(
 	libusb_device *dev);
